Move addDisplayToSettings and canAddDisplay to dashboard utils module
Both function aren't related to the store anymore.
