+++ gsa/src/web/pages/start/dashboard.js
@@ -24,10 +24,9 @@ import React from 'react';
 
 import PropTypes from 'web/utils/proptypes';
 
-import {canAddDisplay} from 'web/store/dashboard/settings/utils';
-
 import {Dashboard} from 'web/components/dashboard/dashboard';
 import {DashboardControls} from 'web/components/dashboard/controls';
+import {canAddDisplay} from 'web/components/dashboard/utils';
 
 import Layout from 'web/components/layout/layout';
