--- gsa/src/web/store/dashboard/settings/utils.js
@@ -1,68 +1,0 @@
-/* Greenbone Security Assistant
- *
- * Authors:
- * Björn Ricks <bjoern.ricks@greenbone.net>
- *
- * Copyright:
- * Copyright (C) 2018 Greenbone Networks GmbH
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 2
- * of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program; if not, write to the Free Software
- * Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA 02110-1301 USA.
- */
-import {isDefined, isArray} from 'gmp/utils/identity';
-
-import {createDisplay, createRow} from 'gmp/commands/dashboards';
-
-export const canAddDisplay = ({rows, maxItemsPerRow, maxRows} = {}) => {
-  if (isArray(rows) && rows.length > 0 &&
-    isDefined(maxItemsPerRow) && isDefined(maxRows)) {
-    const lastRow = rows[rows.length - 1];
-    return lastRow.items.length < maxItemsPerRow || rows.length < maxRows;
-  }
-  return true;
-};
-
-export const addDisplayToSettings = (settings, displayId, uuidFunc) => {
-  const {rows: currentRows = [], maxItemsPerRow} = settings || {};
-
-  const lastRow = isArray(currentRows) && currentRows.length > 0 ?
-    currentRows[currentRows.length - 1] : {items: []};
-
-  const rows = isArray(currentRows) ? [...currentRows] : [];
-  const display = createDisplay(displayId, undefined, uuidFunc);
-
-  let newRow;
-  if (isDefined(maxItemsPerRow) && lastRow.items.length >= maxItemsPerRow) {
-    // create new row
-    newRow = createRow([display], undefined, uuidFunc);
-  }
-  else {
-    // add new display to last row
-    newRow = {
-      ...lastRow,
-      items: [...lastRow.items, display],
-    };
-
-    rows.pop();
-  }
-
-  rows.push(newRow);
-
-  return {
-    ...settings,
-    rows,
-  };
-};
-
-// vim: set ts=2 sw=2 tw=80:
